requirements.txt
# Real-Time Assistive Vision System — runtime dependencies

# Detection & depth
ultralytics>=8.0.0          # YOLOv10 / YOLO model family
torch>=2.0.0                # PyTorch (CPU wheels)
torchvision>=0.15.0         # torchvision (required by MiDaS)
timm>=0.9.0                 # Transformer image models (MiDaS dependency)

# Vision
opencv-python>=4.8.0        # Camera capture & image processing
numpy>=1.21.0               # Numerical operations

# Audio
pyttsx3>=2.90               # Offline text-to-speech

# Config & utilities
PyYAML>=6.0                 # YAML configuration
psutil>=5.9.0               # System / process monitoring

# Optional — better tracker backend (ByteTrack via supervision)
# supervision                # uncomment to enable ByteTrack tracking